Added username view
